Start refactoring GS200 to use SCPI base driver.
