+++ lantz_drivers/yokogawa/model_gs200.py
@@ -12,12 +12,9 @@ from __future__ import (division, unicode_literals, print_function,
                         absolute_import)
 
 from lantz_core.has_features import set_feat
-from lantz_core.action import Action
 from lantz_core.limits import FloatLimitsValidator
 
-from ..standards.iee488 import IEEE488
-from ..standards.dc_sources import DCPowerSource
-
+from ..standards.scpi.dc_sources import SCPIDCPowerSource
 
 VOLTAGE_RESOLUTION = {10e-3: 1e-7,
                       100e-3: 1e-6,
@@ -31,53 +28,25 @@ CURRENT_RESOLUTION = {1e-3: 1e-8,
                       200e-3: 1e-6}
 
 
-# XXXX a large part can be abstracted away in scpi.dc_sources module.
-class YokogawaGS200(IEEE488, DCPowerSource):
+class YokogawaGS200(SCPIDCPowerSource):
     """Driver for the Yokogawa GS200 DC power source.
 
     """
-    function = set_feat(getter=':SOUR:FUNC?',
-                        setter=':SOUR:FUNC {}',
-                        mapping={'Voltage': 'VOLT', 'Current': 'CURR'})
+    voltage = set_feat(limits='voltage')
 
-    output = set_feat(getter='OUTP:STAT?',
-                      setter='OUTP:STAT {}',
-                      mapping={True: 1, False: 0})
-
-    voltage = set_feat(getter=':SOUR:LEV?',
-                       setter=':SOUR:LEV {:E}',
-                       limits='voltage')
-
-    voltage_range = set_feat(getter=':SOUR:RANG?',
-                             setter=':SOUR:RANG {:E}',
-                             values=(10e-3, 100e-3, 1.0, 10.0, 30.0),
+    voltage_range = set_feat(values=(10e-3, 100e-3, 1.0, 10.0, 30.0),
                              discard={'limits': 'voltage'})
 
-    current_limit = set_feat(getter=':SOUR:PROT:CURR?',
-                             setter=':SOUR:PROT:CURR {}',
-                             limits=(1e-3, 200e-3, 1e-3),
+    current_limit = set_feat(limits=(1e-3, 200e-3, 1e-3),
                              checks='{voltage_range} not in (10e-3, 100e-3)\
                                      or {value} == 200e-3')
 
-    current = set_feat(getter=':SOURce:LEVel?',
-                       setter=':SOURce:LEVel {:E}',
-                       limits='current')
+    current = set_feat(limits='current')
 
-    current_range = set_feat(getter=':SOURce:RANGe?',
-                             setter=':SOURce:RANGe {:E}',
-                             values=(1e-3, 10e-3, 100e-3, 200e-3),
+    current_range = set_feat(values=(1e-3, 10e-3, 100e-3, 200e-3),
                              discard={'limits': 'current'})
 
-    voltage_limit = set_feat(getter=':SOUR:PROT:VOLT?',
-                             setter=':SOUR:PROT:VOLT {}',
-                             limits=(1.0, 30.0, 1))
-
-    @Action
-    def read_errors(self):
-        """Read the error message available in the queue.
-
-        """
-        return self.query(':SYST:ERR?')
+    voltage_limit = set_feat(limits=(1.0, 30.0, 1))
 
     def default_check_operation(self, feat, value, i_value, state=None):
         """Check that the operation did not result in any error.
